refactor continue mappingModeler.js
