+++ public/vocables/index.html
@@ -117,6 +117,7 @@
         <script type="module" src="modules/tools/KGcreator/KGcreator_run.js"></script>
         <script type="module" src="modules/tools/mappingModeler/mappingModeler.js"></script>
         <script type="module" src="modules/tools/mappingModeler/tripleFactory.js"></script>
+        <script type="module" src="modules/tools/mappingModeler/mappingTransform.js"></script>
         <script type="module" src="modules/tools/KGcreator/KGcreator_joinTables.js"></script>
 
         <!--   <script type="module" src="modules/tools/KGcreatorGraph.js"></script>-->
