+++ public/vocables/modules/tools/mappingModeler/html/mappingModeler_graphDiv.html
@@ -9,14 +9,8 @@
     <button class="w3-button slsv-button-3" onclick="MappingModeler.classesDialog()"> Details </button>
 
 
-    <!--
-    <button class="w3-button slsv-button-3" onclick="MappingModeler.mappingToKGcreator()">
-        mappingToKGcreator
-    </button>
 
-   -->
-
-    <button class="w3-button slsv-button-3" onclick="MappingModeler.generateBasicMappings()">
+    <button class="w3-button slsv-button-3" onclick="MappingTransform.generateBasicMappings()">
         view mappings
     </button>
